+++ src/index.ts
@@ -1041,7 +1041,7 @@ class Remarkable implements RemarkableApi {
     const [rootEntry, uploadRoot] = await this.raw.putEntries(
       "root",
       rootEntries,
-      schemaVersion,
+      4,
     );
 
     // before updating the root hash, first upload everything
@@ -1128,7 +1128,7 @@ class Remarkable implements RemarkableApi {
     const [rootEntry, uploadRoot] = await this.raw.putEntries(
       "root",
       rootEntries,
-      schemaVersion,
+      4,
     );
 
     // before updating the root hash, first upload everything
@@ -1228,7 +1228,7 @@ class Remarkable implements RemarkableApi {
     const [rootEntry, uploadRoot] = await this.raw.putEntries(
       "root",
       entries,
-      schemaVersion,
+      4,
     );
 
     await Promise.all([uploadEnt, uploadRoot]);
@@ -1314,7 +1314,7 @@ class Remarkable implements RemarkableApi {
     const [rootEntry, uploadRoot] = await this.raw.putEntries(
       "root",
       entries,
-      schemaVersion,
+      4,
     );
 
     await Promise.all([uploadEnt, uploadRoot]);
@@ -1404,7 +1404,7 @@ class Remarkable implements RemarkableApi {
     const [rootEntry, uploadRoot] = await this.raw.putEntries(
       "root",
       newEntries,
-      schemaVersion,
+      4,
     );
     await Promise.all([Promise.all(uploads), uploadRoot]);